Functional messaging framework

+++ common/src/byteBuffer.cpp
@@ -1,12 +1,12 @@
 #include "byteBuffer.h"
 
 
-int ByteBuffer::Length() const
+size_t ByteBuffer::Length() const
 {
     return internal.size();
 }
 
-int ByteBuffer::RemainingBytes() const
+size_t ByteBuffer::RemainingBytes() const
 {
     return Length() - position;
 }
